requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
aws-cdk-lib>=2.89.0, <3.0.0
boto3
constructs>=10.0.5, <11.0.0
poetry
pydantic[dotenv]<=1.10.11
pygit2
